Rename Unit to Feature test directory
