+++ phpunit.xml
@@ -18,8 +18,8 @@
     </php>
 
     <testsuites>
-        <testsuite name="unit">
-            <directory suffix="Test.php">tests/Unit</directory>
+        <testsuite name="feature">
+            <directory suffix="Test.php">tests/Feature</directory>
         </testsuite>
     </testsuites>
